Moved Section from `guui` to `frontend`
