+++ packages/frontend/web/pages/Interview.tsx
@@ -8,7 +8,7 @@ import { SubNav } from '@frontend/web/components/Header/Nav/SubNav/SubNav';
 import { CookieBanner } from '@frontend/web/components/CookieBanner';
 import { MobileStickyContainer } from '@frontend/web/components/AdSlot';
 import { OutbrainContainer } from '@frontend/web/components/Outbrain';
-import { Section } from '@guardian/guui';
+import { Section } from '@frontend/web/components/Section';
 
 export const Interview: React.FC<{
     data: ArticleProps;
