+++ packages/guui/index.ts
@@ -2,6 +2,5 @@
 export { CloseButton } from './components/CloseButton/CloseButton';
 export { Dropdown } from './components/Dropdown/Dropdown';
 export { BigNumber } from './components/BigNumber/BigNumber';
-export { Section } from './components/Section/Section';
 export { Cols as GridCols, Row as GridRow } from './components/Grid/Grid';
 export { Byline } from './components/Byline/Byline';
